refactor: extract feedback function to a seperate file
